Added link to google maps
Just in case we don't have enough time to create coding for api request and display for a map, which it pretty much seems like we won't, I just linked the map button on our main page to open up google maps in another window when the button is clicked. Hopefully they will let us do this if we don't have enough time since we are down a partner.