Updated index.html
Sorry, but your changes made all of the JavaScript functions inoperable and I spent an hour or so trying to make it work and we just don't have enough time to tweak the coding to make it at least a slightly functioning website. I at least tried to mimic the style you had going in the change though.